Major plugin structure change
All hooks are now in the Hooks namespace, which will invoke other parts of the
plugin. This will make the plugin more bite-size chunks, adding to the
ease-of-use and separates responsibilities, instead of resulting in a superclass
in the Plugin class.

This will also free the way for more functionality in the plugin.