UTF-8 fix for DOMDocument, remove wp: comments
This removes the Gutenberg comments from the filtered response, since we don't
actually want to show that on the front-end to users.

It also solves a bug causing PHP's DOMDocument to read the user's post content
as ISO-8859-1, instead of UTF-8. This is solved by adding an extra <?xml>
statement to the file.